worker: ensure firmware install is ordered, this was left out in a refactor
